requirements.txt
fastapi
uvicorn
curl-cffi
pandas
playwright
opencv-python-headless
numpy
google-cloud-bigquery
google-cloud-tasks
pydantic
python-dotenv
pytz